cascade_mcp_pack/research/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cascade-research-mcp"
version = "0.1.0"
description = "Thin MCP wrapper package for CASCADE research_server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crawl4ai>=0.7.3",
    "mcp>=1.7.1",
    "supabase>=2.18.1",
    "openai>=1.81.0",
    "python-dotenv>=1.0.1",
    "neo4j>=5.28.2",
]

[project.optional-dependencies]
test = ["pytest>=8.0.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
